requirements.txt
Django==3.2.15
Pillow==9.2.0

django-allauth==0.51.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0

flake8==5.0.4
isort==5.10.1

stripe==4.1.0

redis==4.3.4
celery==5.2.7
django-redis==5.2.0

psycopg2-binary==2.9.3
